# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Artemis Application Server API
  contact: {name: Stephan Krusche, email: krusche@tum.de}
  version: "10.0"
servers:
- {url: 'http://localhost:8080', description: Generated server url}
...
paths:
...
  /api/hyperion/programming-exercises/{exerciseId}/generation/runs/{runId}/revert:
    post:
      tags: [hyperion-exercise-generation]
      operationId: revertExerciseGeneration
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      - name: runId
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExerciseGenerationRevertResult'}
  /api/hyperion/programming-exercises/{exerciseId}/generate-exercise:
    post:
      tags: [hyperion-exercise-generation]
      operationId: generateExercise
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExerciseGenerationRequest'}
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExerciseGenerationJobStart'}
...
  /api/hyperion/courses/{courseId}/programming-exercises/generation/metadata-suggestion:
    post:
      tags: [hyperion-exercise-generation]
      operationId: suggestGenerationMetadata
      parameters:
      - name: courseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExerciseGenerationMetadataSuggestionRequest'}
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExerciseGenerationMetadataSuggestionResponse'}
...
  /api/hyperion/programming-exercises/{exerciseId}/generation/capabilities:
    get:
      tags: [hyperion-exercise-generation]
      operationId: getGenerationCapabilities
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExerciseGenerationCapabilities'}
  /api/hyperion/programming-exercises/{exerciseId}/generate-exercise/status:
    get:
      tags: [hyperion-exercise-generation]
      operationId: getExerciseGenerationStatus
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExerciseGenerationStatus'}
  /api/hyperion/programming-exercises/{exerciseId}/generate-exercise/artifacts:
    get:
      tags: [hyperion-exercise-generation]
      operationId: getRetainedGenerationArtifacts
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExerciseGenerationRetainedArtifacts'}
...
  /api/hyperion/programming-exercises/{exerciseId}/generate-exercise/jobs/{jobId}:
    delete:
      tags: [hyperion-exercise-generation]
      operationId: cancelExerciseGeneration
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      - name: jobId
        in: path
        required: true
        schema: {type: string}
      responses:
        '200': {description: OK}
...
components:
  schemas:
...
    ExerciseGenerationFileChange:
      type: object
      description: A file changed by the generation agent
      properties:
        type:
          type: string
          description: Constant discriminator identifying a file change on the shared
            topic
          enum: [FILE_CHANGE]
        path: {type: string, description: Workspace-relative file path}
        repo:
          type: string
          description: Owning repository bucket
          enum: [solution, template, tests, other]
        action:
          type: string
          description: The successful file tool operation
          enum: [write, edit, delete]
        turn: {type: integer, format: int32, description: The agent turn the change
            happened on}
        timestamp: {type: string, format: date-time, description: The moment the change
            was produced}
      required: [action, path, repo, timestamp, turn, type]
...
    ExerciseGenerationMetadataSuggestionResponse:
      type: object
      description: The metadata Artemis derives for an exercise it is about to generate
      properties:
        title: {type: string}
        shortName: {type: string}
        packageName: {type: string}
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        maxPoints: {type: number, format: double}
      required: [difficulty, maxPoints, packageName, shortName, title]
...
    ExerciseGenerationRetainedFile:
      type: object
      description: A file of a generated candidate that was not saved to the exercise
      properties:
        repo:
          type: string
          description: Owning repository bucket
          enum: [solution, template, tests]
        path: {type: string, description: Repository-relative file path}
        content: {type: string, description: The file's text content as the agent
            left it}
      required: [content, path, repo]
...
    ExerciseGenerationUsage:
      type: object
      properties:
        modelCalls: {type: integer, format: int64}
        toolCalls: {type: integer, format: int64}
        agentTurns: {type: integer, format: int64}
        attempts: {type: integer, format: int64}
        inputTokens: {type: integer, format: int64}
        outputTokens: {type: integer, format: int64}
        cachedInputTokens: {type: integer, format: int64}
        cachedInputTokensComplete: {type: boolean}
        estimatedCostEur: {type: number, format: double}
        estimatedCostEurComplete: {type: boolean}
        models:
          type: array
          items: {type: string}
        providerRequestIds:
          type: array
          items: {type: string}
        providerRequestIdsComplete: {type: boolean}
      required: [agentTurns, attempts, cachedInputTokens, cachedInputTokensComplete,
        estimatedCostEur, estimatedCostEurComplete, inputTokens, modelCalls, models,
        outputTokens, providerRequestIds, providerRequestIdsComplete, toolCalls]
    ExerciseGenerationJobStart:
      type: object
      properties:
        jobId: {type: string}
        exerciseId: {type: integer, format: int64}
        sourceExerciseId: {type: integer, format: int64}
      required: [jobId]
...
    ExerciseGenerationEvent:
      type: object
      description: A progress event streamed to the instructor while an agentic whole-exercise
        generation or adaptation runs
      properties:
        type:
          type: string
          description: The event kind
          enum: [STARTED, PROGRESS, DONE, CANCELLED, ERROR]
        message: {type: string, description: Human-readable progress or result message}
        phase:
          type: string
          description: Stable instructor-facing phase of the generation journey
          enum: [PREPARING, DESIGNING, VERIFYING, REVIEWING, REPAIRING, SAVING]
        completionStatus:
          type: string
          description: On a terminal DONE event, whether the run succeeded, needs
            review, or partially completed
          enum: [SUCCESS, NEEDS_REVIEW, PARTIAL]
        verdict: {$ref: '#/components/schemas/ExerciseGenerationVerdict', description: 'On
            a terminal event, the structured verification verdict'}
        liveExerciseChanged: {type: boolean, description: 'On a terminal event, whether
            the live exercise changed and open editors should refresh'}
        savedRepositoryCommits:
          type: object
          additionalProperties: {type: string}
          description: Exact saved commit hashes keyed by repository name
        savedExerciseVersionId: {type: integer, format: int64, description: 'The exact
            saved exercise version id, on a successful terminal event'}
        terminationReason:
          type: string
          description: On a terminal event, why the generation run ended
          enum: [CONVERGED, REPAIR_BUDGET_EXHAUSTED, ATTEMPT_CAP_REACHED, MECHANICAL_REPAIR_EXHAUSTED,
            POST_REPAIR_CORRECTION_EXHAUSTED, NO_SCHEDULABLE_SURFACE, REVIEW_UNAVAILABLE,
            UNCHANGED_CANDIDATE_RESUBMITTED, REPAIR_DID_NOT_IMPROVE, NO_ADMISSIBLE_CONCEPT,
            CONCEPT_ADMITTED_WITH_FINDINGS, CANCELLED, DEADLINE_EXCEEDED, TOKEN_BUDGET_EXHAUSTED,
            AGENT_ERROR, ENVIRONMENT_UNAVAILABLE, RUN_FAILED, NOT_STARTED]
        repairRound: {$ref: '#/components/schemas/ExerciseGenerationRepairRound',
          description: 'On a repair-round progress event, that round''s finding bookkeeping'}
        activity: {$ref: '#/components/schemas/ExerciseGenerationActivity', description: 'On
            a progress event produced from inside the agent loop, what the run was
            doing at that moment'}
        liveUsage: {$ref: '#/components/schemas/ExerciseGenerationLiveUsage', description: 'On
            an activity or phase event, what the run has spent up to that moment'}
        timestamp: {type: string, format: date-time, description: The moment the event
            was produced}
      required: [timestamp, type]
...
    ExerciseGenerationStatus:
      type: object
      properties:
        jobId: {type: string}
        running: {type: boolean}
        mode:
          type: string
          enum: [GENERATE, ADAPT]
        events:
          type: array
          items: {$ref: '#/components/schemas/ExerciseGenerationEvent'}
        fileChanges:
          type: array
          items: {$ref: '#/components/schemas/ExerciseGenerationFileChange'}
        revertAvailable: {type: boolean}
        revertJobId: {type: string}
        revertMode:
          type: string
          enum: [GENERATE, ADAPT]
        ownedByCaller: {type: boolean}
        cancellable: {type: boolean}
        specDocument: {type: string}
        usage: {$ref: '#/components/schemas/ExerciseGenerationUsage'}
        accountingState:
          type: string
          description: Whether the reported usage is a complete account of a generation
            run's provider spend
          enum: [PENDING, COMPLETE, INCOMPLETE]
        effortProfile: {type: string}
        artifactsRetained: {type: boolean, description: Whether a current or retained
            candidate snapshot from this run is readable}
        input: {$ref: '#/components/schemas/ExerciseGenerationInput'}
        run: {$ref: '#/components/schemas/AuthoringRun'}
      required: [accountingState, artifactsRetained, cancellable, events, fileChanges,
        jobId, ownedByCaller, revertAvailable, running]
...
    ExerciseGenerationMetadataSuggestionRequest:
      type: object
      description: Request for the derived metadata of an exercise an instructor is
        about to generate
      properties:
        prompt: {type: string, description: The instructor's brief for the exercise
            to be generated, maxLength: 8000, minLength: 0}
        projectType:
          type: string
          description: The project type the exercise will use
          enum: [MAVEN_MAVEN, PLAIN_MAVEN, PLAIN, XCODE, FACT, GCC, PLAIN_GRADLE,
            GRADLE_GRADLE, MAVEN_BLACKBOX]
      required: [prompt]
...
    ExerciseGenerationRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [GENERATE, ADAPT]
        prompt: {type: string, maxLength: 8000, minLength: 0}
        selectedFeedbackThreadIds:
          type: array
          items: {type: integer, format: int64}
          maxItems: 25
          minItems: 0
        effortProfile: {type: string, maxLength: 64, minLength: 0}
        maxTokens: {type: integer, format: int64, exclusiveMinimum: 0}
        maxJobDuration: {type: string, format: duration}
      required: [mode]
...
    ExerciseGenerationRevertResult:
      type: object
      description: Result of reverting the last Hyperion generation or adaptation
      properties:
        fullyReverted: {type: boolean, description: Whether every captured repository
            was reverted}
        revertedRepositories:
          type: array
          description: Repositories that were reset
          items: {type: string}
        completedAt: {type: string, format: date-time, description: Server time after
            the revert completed}
      required: [completedAt, fullyReverted, revertedRepositories]
...
    ExerciseGenerationRetainedArtifacts:
      type: object
      description: The current bounded, read-only generated candidate, also retained
        for inspection when an unsaved run ends
      properties:
        jobId: {type: string}
        completeness:
          type: string
          description: Whether the retained files are a complete account of what the
            run produced
          enum: [COMPLETE, PARTIAL]
        problemStatement: {type: string}
        specDocument: {type: string}
        files:
          type: array
          items: {$ref: '#/components/schemas/ExerciseGenerationRetainedFile'}
        empty: {type: boolean}
      required: [completeness, files, jobId]
...
    ExerciseGenerationFeedback:
      type: object
      properties:
        targetType: {type: string}
        filePath: {type: string}
        lineNumber: {type: integer, format: int32}
        comments:
          type: array
          items: {type: string}
...
    ExerciseGenerationCapabilities:
      type: object
      properties:
        supported: {type: boolean}
        canGenerate: {type: boolean}
        canAdapt: {type: boolean}
        canCreateVariant: {type: boolean}
        busy: {type: boolean}
        capacityAvailable: {type: boolean}
        restriction: {type: string}
...
    ExerciseGenerationInput:
      type: object
      properties:
        prompt: {type: string}
        reviewFeedback:
          type: array
          items: {$ref: '#/components/schemas/ExerciseGenerationFeedback'}
        sourceExerciseId: {type: integer, format: int64}
...