# src/main/java/de/tum/cit/aet/artemis/hyperion/web/HyperionExerciseGenerationResource.java
package de.tum.cit.aet.artemis.hyperion.web;

import java.time.Instant;
import java.util.List;
import java.util.Optional;

import jakarta.validation.Valid;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.context.annotation.Conditional;
import org.springframework.context.annotation.Lazy;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import de.tum.cit.aet.artemis.account.domain.User;
import de.tum.cit.aet.artemis.account.repository.UserRepository;
import de.tum.cit.aet.artemis.core.exception.EntityNotFoundException;
import de.tum.cit.aet.artemis.core.security.annotations.EnforceAtLeastEditor;
import de.tum.cit.aet.artemis.core.security.annotations.enforceRoleInCourse.EnforceAtLeastEditorInCourse;
import de.tum.cit.aet.artemis.core.security.annotations.enforceRoleInExercise.EnforceAtLeastEditorInExercise;
import de.tum.cit.aet.artemis.core.service.featureusage.FeatureUsage;
import de.tum.cit.aet.artemis.hyperion.config.HyperionExerciseGenerationEnabled;
import de.tum.cit.aet.artemis.hyperion.dto.ExerciseGenerationCapabilitiesDTO;
import de.tum.cit.aet.artemis.hyperion.dto.ExerciseGenerationEffortProfileDTO;
import de.tum.cit.aet.artemis.hyperion.dto.ExerciseGenerationFileChangeDTO;
import de.tum.cit.aet.artemis.hyperion.dto.ExerciseGenerationJobStartDTO;
import de.tum.cit.aet.artemis.hyperion.dto.ExerciseGenerationMetadataSuggestionRequestDTO;
import de.tum.cit.aet.artemis.hyperion.dto.ExerciseGenerationMetadataSuggestionResponseDTO;
import de.tum.cit.aet.artemis.hyperion.dto.ExerciseGenerationRequestDTO;
import de.tum.cit.aet.artemis.hyperion.dto.ExerciseGenerationRetainedArtifactsDTO;
import de.tum.cit.aet.artemis.hyperion.dto.ExerciseGenerationRevertResultDTO;
import de.tum.cit.aet.artemis.hyperion.dto.ExerciseGenerationStatusDTO;
import de.tum.cit.aet.artemis.hyperion.dto.GenerationMode;
import de.tum.cit.aet.artemis.hyperion.service.HyperionExerciseMetadataSuggestionService;
import de.tum.cit.aet.artemis.hyperion.service.exercisegeneration.orchestration.GenerationAdmissionService;
import de.tum.cit.aet.artemis.hyperion.service.exercisegeneration.orchestration.GenerationJobService;
import de.tum.cit.aet.artemis.hyperion.service.exercisegeneration.orchestration.HyperionEffortProfileService;
import de.tum.cit.aet.artemis.hyperion.service.exercisegeneration.persistence.ExerciseGenerationRevertService;
import de.tum.cit.aet.artemis.hyperion.service.exercisegeneration.profile.GenerationCapabilityService;
import de.tum.cit.aet.artemis.hyperion.service.exercisegeneration.profile.GenerationRequestService;
import de.tum.cit.aet.artemis.programming.domain.ProgrammingExercise;
import de.tum.cit.aet.artemis.programming.domain.ProgrammingLanguage;
import de.tum.cit.aet.artemis.programming.domain.RepositoryType;
import de.tum.cit.aet.artemis.programming.repository.ProgrammingExerciseRepository;

/**
 * REST controller for Hyperion's agentic whole-exercise generation and adaptation.
 * <p>
 * One endpoint drives both {@link GenerationMode} values. Mechanically verified work is saved to the exercise and versioned, and automated quality findings become review comments
 * for the instructor; mechanically invalid or interrupted candidates are not saved. A run is a multi-minute async job addressed by the returned {@code jobId}, streaming progress
 * over the websocket topic {@code /user/topic/hyperion/exercise-generation/jobs/{jobId}}.
 */
@Conditional(HyperionExerciseGenerationEnabled.class)
@Lazy
@RestController
@FeatureUsage("authoring-assistance/exercise-generation")
@RequestMapping("api/hyperion/")
public class HyperionExerciseGenerationResource {

    private static final Logger log = LoggerFactory.getLogger(HyperionExerciseGenerationResource.class);

    private static final String ENTITY_NAME = "hyperionExerciseGeneration";

    private final GenerationAdmissionService admission;

    private final UserRepository userRepository;

    private final ProgrammingExerciseRepository programmingExerciseRepository;

    private final GenerationCapabilityService capabilities;

    private final GenerationJobService jobService;

    private final GenerationRequestService generationRequestService;

    private final ExerciseGenerationRevertService generationRevertService;

    private final HyperionEffortProfileService effortProfileService;

    private final HyperionExerciseMetadataSuggestionService metadataSuggestionService;

    public HyperionExerciseGenerationResource(UserRepository userRepository, ProgrammingExerciseRepository programmingExerciseRepository, GenerationCapabilityService capabilities,
            GenerationJobService jobService, GenerationRequestService generationRequestService, GenerationAdmissionService admission,
            ExerciseGenerationRevertService generationRevertService, HyperionEffortProfileService effortProfileService,
            HyperionExerciseMetadataSuggestionService metadataSuggestionService) {
        this.admission = admission;
        this.userRepository = userRepository;
        this.programmingExerciseRepository = programmingExerciseRepository;
        this.capabilities = capabilities;
        this.jobService = jobService;
        this.generationRequestService = generationRequestService;
        this.generationRevertService = generationRevertService;
        this.effortProfileService = effortProfileService;
        this.metadataSuggestionService = metadataSuggestionService;
    }

    /**
     * POST courses/{courseId}/programming-exercises/generation/metadata-suggestion : derives the metadata of the exercise the instructor's brief describes — its title,
     * difficulty, short name, package name, and points.
     * <p>
     * Course-scoped rather than exercise-scoped because it is asked before the exercise exists, and titles and short names have to be free in the course it will be created in. It
     * answers with values Artemis will accept even when the model is unavailable, so a failed suggestion never stands between the instructor and generation.
     *
     * @param courseId the course the exercise will be created in
     * @param request  the instructor's brief and the project type the exercise will use
     * @return 200 with the derived metadata
     */
    @PostMapping("courses/{courseId}/programming-exercises/generation/metadata-suggestion")
    @EnforceAtLeastEditorInCourse
    public ResponseEntity<ExerciseGenerationMetadataSuggestionResponseDTO> suggestGenerationMetadata(@PathVariable long courseId,
            @Valid @RequestBody ExerciseGenerationMetadataSuggestionRequestDTO request) {
        log.debug("REST request to suggest the generation metadata for course [{}]", courseId);
        return ResponseEntity.ok(metadataSuggestionService.suggestMetadata(courseId, request.prompt(), request.projectType()));
    }

    /**
     * POST programming-exercises/{exerciseId}/generate-exercise : starts an agentic whole-exercise generation/adaptation run in the request's explicit mode.
     *
     * @param exerciseId the programming exercise id
     * @param request    the request holding the explicit mode and the optional prompt / selected feedback threads
     * @return 202 Accepted with the started job id the client uses to reattach and stream progress; 409 if a run is already active for the exercise; 503 if no build agent
     *         currently has a free sandbox slot
     */
    @PostMapping("programming-exercises/{exerciseId}/generate-exercise")
    @EnforceAtLeastEditorInExercise
    public ResponseEntity<ExerciseGenerationJobStartDTO> generateExercise(@PathVariable long exerciseId, @Valid @RequestBody ExerciseGenerationRequestDTO request) {
        log.debug("REST request to run agentic exercise generation ({}) for exercise [{}]", request.mode(), exerciseId);
        return ResponseEntity.accepted()
                .body(new ExerciseGenerationJobStartDTO(admission.start(userRepository.getUserWithAuthorities(), exerciseId, request), exerciseId, exerciseId));
    }

    /**
     * Reads the same exercise-specific policy used at admission. Unsupported actions can be hidden while busy workers remain a retryable state.
     *
     * @param exerciseId exercise whose authoring actions are requested
     * @return supported actions and current availability
     */
    @GetMapping("programming-exercises/{exerciseId}/generation/capabilities")
    @EnforceAtLeastEditorInExercise
    public ResponseEntity<ExerciseGenerationCapabilitiesDTO> getGenerationCapabilities(@PathVariable long exerciseId) {
        return ResponseEntity.ok(capabilities.describe(loadExercise(exerciseId)));
    }

    /**
     * GET programming-exercises/generation/supported-languages : the languages whole-exercise generation can be verified for, so clients fetch the set instead of hardcoding it.
     * <p>
     * Not exercise-scoped, so it is guarded by the least-privileged global role that can create exercises ({@link EnforceAtLeastEditor}) rather than by course membership.
     *
     * @return the supported programming languages
     */
    @GetMapping("programming-exercises/generation/supported-languages")
    @EnforceAtLeastEditor
    public ResponseEntity<List<ProgrammingLanguage>> getSupportedGenerationLanguages() {
        log.debug("REST request to get the Hyperion generation-supported programming languages");
        return ResponseEntity.ok(generationRequestService.supportedGenerationLanguages().stream().sorted().toList());
    }

    /**
     * GET programming-exercises/generation/effort-profiles : the effort profiles an instructor may pick for a run. Empty when the deployment configures none, in which case every
     * run uses the deployment-wide configuration.
     * <p>
     * Name and label only: model ids and budgets are admin-owned and must not be exposed here.
     * <p>
     * Not exercise-scoped, so it is guarded by the least-privileged global role that can create exercises ({@link EnforceAtLeastEditor}), mirroring
     * {@link #getSupportedGenerationLanguages()}.
     *
     * @return the selectable effort profiles, in configured order
     */
    @GetMapping("programming-exercises/generation/effort-profiles")
    @EnforceAtLeastEditor
    public ResponseEntity<List<ExerciseGenerationEffortProfileDTO>> getGenerationEffortProfiles() {
        log.debug("REST request to get the Hyperion generation effort profiles");
        return ResponseEntity.ok(effortProfileService.selectableProfiles());
    }

    /**
     * GET programming-exercises/{exerciseId}/generate-exercise/status : returns the caller's current or most-recent run for the exercise (id, whether it is still running, and the
     * transcript so far), so a client that (re)loads the page can replay the progress and reattach to a live run. Returns 204 when there is nothing to show.
     *
     * @param exerciseId the programming exercise id
     * @return the run status with the replayable transcript, or 204 if none is retained for the caller
     */
    @GetMapping("programming-exercises/{exerciseId}/generate-exercise/status")
    @EnforceAtLeastEditorInExercise
    public ResponseEntity<ExerciseGenerationStatusDTO> getExerciseGenerationStatus(@PathVariable long exerciseId) {
        log.debug("REST request to get the agentic exercise generation status for exercise [{}]", exerciseId);
        ProgrammingExercise exercise = loadExercise(exerciseId);
        User user = userRepository.getUserWithAuthorities();
        Optional<ExerciseGenerationRevertService.RevertibleRun> revertibleRun = generationRevertService.findRevertibleRun(exerciseId).filter(run -> canOfferRevert(exercise));
        Optional<ExerciseGenerationStatusDTO> retainedStatus = jobService.getStatus(user, exercise);
        if (retainedStatus.isPresent()) {
            ExerciseGenerationStatusDTO status = retainedStatus.get();
            return ResponseEntity.ok(new ExerciseGenerationStatusDTO(status.jobId(), status.running(), status.mode(), status.events(), status.fileChanges(),
                    revertibleRun.isPresent(), revertibleRun.map(ExerciseGenerationRevertService.RevertibleRun::jobId).orElse(null),
                    revertibleRun.map(ExerciseGenerationRevertService.RevertibleRun::mode).orElse(null), status.ownedByCaller(), status.cancellable(), status.specDocument(),
                    status.usage(), status.accountingState(), status.effortProfile(), status.artifactsRetained(), status.input()));
        }
        return revertibleRun.<ResponseEntity<ExerciseGenerationStatusDTO>>map(run -> ResponseEntity.ok(ExerciseGenerationStatusDTO.revertOnly(run.jobId(), run.mode())))
                .orElseGet(() -> ResponseEntity.noContent().build());
    }

    /**
     * GET programming-exercises/{exerciseId}/generate-exercise/artifacts : returns the current candidate of the caller's most recent run. It is updated after each successful
     * structured file mutation and remains available when an unsaved run ends. Read-only, bounded, and held only for the lifetime of the run's replay evidence.
     * <p>
     * Owner-only: editor rights in the course are necessary but not sufficient, because this returns the verbatim content of an unreviewed draft rather than a progress narrative.
     *
     * @param exerciseId the programming exercise id
     * @return the current or retained candidate, or 404 before the run writes anything or after its evidence expires
     */
    @GetMapping("programming-exercises/{exerciseId}/generate-exercise/artifacts")
    @EnforceAtLeastEditorInExercise
    public ResponseEntity<ExerciseGenerationRetainedArtifactsDTO> getRetainedGenerationArtifacts(@PathVariable long exerciseId) {
        log.debug("REST request to get the current generation candidate for exercise [{}]", exerciseId);
        ProgrammingExercise exercise = loadExercise(exerciseId);
        User user = userRepository.getUserWithAuthorities();
        return jobService.getRetainedArtifacts(user, exercise).map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
    }

    /**
     * DELETE programming-exercises/{exerciseId}/generate-exercise/jobs/{jobId} : cancels a running generation job and immediately exposes its terminal status. An in-flight
     * provider request may still finish at the transport layer, but its response is fenced from tool execution and completion publication.
     * <p>
     * Editor rights in the course are necessary but not sufficient: job ids are observable, so only the user who started the job may cancel it.
     *
     * @param exerciseId the programming exercise id
     * @param jobId      the job id to cancel
     * @return 200 if a matching active job owned by the caller was marked for cancellation, 404 otherwise
     */
    @DeleteMapping("programming-exercises/{exerciseId}/generate-exercise/jobs/{jobId}")
    @EnforceAtLeastEditorInExercise
    public ResponseEntity<Void> cancelExerciseGeneration(@PathVariable long exerciseId, @PathVariable String jobId) {
        log.debug("REST request to cancel agentic exercise generation job [{}] for exercise [{}]", jobId, exerciseId);
        User user = userRepository.getUserWithAuthorities();
        boolean cancelled = jobService.requestCancellation(exerciseId, jobId, user);
        return cancelled ? ResponseEntity.ok().build() : ResponseEntity.notFound().build();
    }

    /**
     * POST programming-exercises/{exerciseId}/generation/runs/{runId}/revert : restores the selected saved run only if it is still the latest mutation, resetting its repositories
     * back to the state captured before persistence.
     *
     * @param exerciseId the programming exercise id
     * @param runId      the exact run selected by the instructor
     * @return 200 if a baseline was found and fully reverted; 409 if at least one repository failed to revert; 404 if there is nothing to revert
     */
    @PostMapping("programming-exercises/{exerciseId}/generation/runs/{runId}/revert")
    @EnforceAtLeastEditorInExercise
    public ResponseEntity<ExerciseGenerationRevertResultDTO> revertExerciseGeneration(@PathVariable long exerciseId, @PathVariable String runId) {
        log.debug("REST request to revert the last saved agentic generation run of exercise [{}]", exerciseId);
        ProgrammingExercise exercise = loadExercise(exerciseId);
        validateDraftExercise(exercise);
        User user = userRepository.getUserWithAuthorities();
        String revertSlot = jobService.claimRevertSlot(user, exerciseId);
        var consistent = new java.util.concurrent.atomic.AtomicBoolean();
        try {
            Optional<String> revertibleJobId = generationRevertService.findRevertibleJobId(exerciseId);
            if (revertibleJobId.isEmpty() || !runId.equals(revertibleJobId.get())) {
                consistent.set(!jobService.isRevertRecoveryRetry(revertSlot));
                return revertibleJobId.isEmpty() ? ResponseEntity.notFound().build() : ResponseEntity.status(HttpStatus.CONFLICT).build();
            }
            return generationRevertService
                    .revert(exercise, user,
                            () -> jobService.isOwnedActiveJob(exerciseId, revertSlot) && programmingExerciseRepository.isUnreleasedAndWithoutStudentParticipations(exerciseId))
                    .map(result -> {
                        if (result.fullyReverted()) {
                            consistent.set(true);
                            revertibleJobId.ifPresent(jobId -> jobService.discardRetainedRun(exerciseId, jobId));
                        }
                        else if (!result.mutationAttempted()) {
                            consistent.set(!jobService.isRevertRecoveryRetry(revertSlot));
                        }
                        ExerciseGenerationRevertResultDTO body = new ExerciseGenerationRevertResultDTO(result.fullyReverted(),
                                result.revertedRepositories().stream().map(HyperionExerciseGenerationResource::repositoryLabel).toList(), Instant.now());
                        return result.fullyReverted() ? ResponseEntity.ok(body) : ResponseEntity.status(HttpStatus.CONFLICT).body(body);
                    }).orElseGet(() -> {
                        consistent.set(!jobService.isRevertRecoveryRetry(revertSlot));
                        return ResponseEntity.notFound().build();
                    });
        }
        finally {
            if (consistent.get()) {
                jobService.clearRevertSlot(exerciseId, revertSlot);
            }
            else {
                jobService.retainRevertRecoverySlot(exerciseId, revertSlot);
            }
        }
    }

    private ProgrammingExercise loadExercise(long exerciseId) {
        ProgrammingExercise exercise = programmingExerciseRepository.findWithAllParticipationsById(exerciseId)
                .orElseThrow(() -> new EntityNotFoundException("Programming Exercise", exerciseId));
        return exercise;
    }

    /**
     * Reuses the labels of {@link ExerciseGenerationFileChangeDTO} so both endpoints name the same three repositories identically and share the client's translation keys.
     * {@link RepositoryType#getName()} cannot be used: it returns the LocalVC-facing {@code "exercise"} for {@link RepositoryType#TEMPLATE}.
     */
    private static String repositoryLabel(RepositoryType repositoryType) {
        return switch (repositoryType) {
            case TEMPLATE -> ExerciseGenerationFileChangeDTO.REPOSITORY_TEMPLATE;
            case SOLUTION -> ExerciseGenerationFileChangeDTO.REPOSITORY_SOLUTION;
            case TESTS -> ExerciseGenerationFileChangeDTO.REPOSITORY_TESTS;
            default -> repositoryType.getName();
        };
    }

    /**
     * Hyperion writes directly to the exercise repositories, so it must only run on exercises no student has started. An exercise whose release date lies in the past is
     * rejected; one without a release date is accepted, because instructors leave the date empty while authoring and {@link ProgrammingExercise#isReleased()} would treat that as
     * released. An exercise that is visible but has no participations stays eligible: the participation endpoints refuse to start one while a run holds the exercise, so a
     * student cannot copy a template that is about to change.
     */
    private void validateDraftExercise(ProgrammingExercise exercise) {
        capabilities.requireMutable(exercise);
    }

    private boolean canOfferRevert(ProgrammingExercise exercise) {
        return capabilities.mutationRestriction(exercise) == null && (!jobService.hasActiveJob(exercise.getId()) || jobService.isRevertRecoveryPending(exercise.getId()));
    }

}